# Render immediately when typing