@Tyriar 2022-10-21
Managed to get typometer working, I needed to uncheck Native API. Here are the initial results:

<img width="395" alt="image" src="https://user-images.githubusercontent.com/2193314/197224323-97fbcff2-2d21-4dbf-b273-6128157f9a77.png">