@Tyriar 2022-10-21
With electron textarea:

<img width="296" alt="image" src="https://user-images.githubusercontent.com/2193314/197240428-199434fa-4a4b-4ebb-a50c-c29597238a9c.png">