@tfriedel 2022-10-21
On which Windows version do you get these results? I get similar results on Windows 10, but not on Windows 11. Since Windows 11 launched I repeatedly ran experiments with typometer, the last time 3 months ago. On two separate computers, a laptop and my desktop, I can't even get notepad to a latency of 15-18 with Windows 11, which is why I'm still using Windows 10.