@Tyriar 2022-10-21
@tfriedel I'm on Windows 11, does your monitor have a high refresh rate and do you have a fast CPU?