@tfriedel 2022-10-21
@Tyriar The monitor (external) and laptop display  both have 60 hz. Also both have Nvidia GPUs. CPUs are pretty fast (i5/i7). But it's good to know that apparently it's possible to get low latency on win 11 even in notepad. May give it another shot then.
The only exception where I did measure low latency were browsers like chrome.