@tfriedel 2022-10-22
I know this is likely not the correct thread, but I think the general issue with Win11 might be interesting for you.
quick update: I tried win11, newest insider build on my laptop again and the measured latency for notepad is 50+ ms, for vs code (regular, not insiders build) it's 35ms. Really no idea why! I was thinking the dwm in Win11 might be different and could introduce some additional buffering?