@alexdima 2022-10-22
We will abandon this exploration because, even if the DOM is updated before the animation frame, it is not rendered sooner by the browser.